refactpr: ReceiverModal 커스텀 훅, 스키마 파일 분리
